+++ unstructured/ingest/v2/processes/embedder.py
@@ -58,6 +58,13 @@ class EmbedderConfig(EnhancedDataClassJsonMixin):
             )
 
             return VertexAIEmbeddingEncoder(config=VertexAIEmbeddingConfig(**kwargs))
+        elif self.embedding_provider == "mixedbread-ai":
+            from unstructured.embed.mixedbreadai import (
+                MixedbreadAIEmbeddingConfig,
+                MixedbreadAIEmbeddingEncoder,
+            )
+
+            return MixedbreadAIEmbeddingEncoder(config=MixedbreadAIEmbeddingConfig(**kwargs))
         else:
             raise ValueError(f"{self.embedding_provider} not a recognized encoder")